feat: calls offloader deployment
